LPS-66913 Move CompoundSessionIdFilter to module
