+++ modules/apps/foundation/portal/portal-compound-session-id/src/main/java/com/liferay/portal/compound/session/id/CompoundSessionIdServletRequestFactory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.servlet.filters.compoundsessionid;
+package com.liferay.portal.compound.session.id;
 
 import javax.servlet.http.HttpServletRequest;
